split tui modules for maintainability
