+++ docs/technical-details.md
@@ -43,9 +43,10 @@ The session manager adds a per-request backend target on top of the global defau
 The TUI is intentionally separated from backend internals:
 
 - `src/backend_client.rs` is the reusable client boundary. It knows socket paths, JSON request/response wrapping, terminal frame IO, and error types. It does not render UI.
-- `src/tui.rs` owns TUI domain models, snapshot parsing, selection state, key mapping, text snapshot rendering, and Ratatui widgets. It consumes `BackendClient` and backend JSON, but does not know `BuiltinBackendInner` internals.
+- `src/tui.rs` owns TUI application state, refresh flow, selection, text snapshot rendering, and terminal attach/input orchestration. It consumes `BackendClient` and backend JSON, but does not know `BuiltinBackendInner` internals.
+- `src/tui_model.rs` owns TUI DTOs and backend snapshot parsing. `src/tui_input.rs` maps keys to terminal bytes and menu shortcuts. `src/tui_render.rs` owns Ratatui widget rendering. `src/tui_tests.rs` keeps TUI app/render regression coverage out of the production module.
 - `src/bin/herdr-webui-tui.rs` is a thin binary shell for CLI parsing, terminal raw mode, the event loop, and the live terminal reader/writer thread.
-- Terminal text parsing is split by responsibility. `src/terminal_text.rs` provides shared plain-text terminal rewrite handling for backend snapshots. `src/tui.rs` adds styled ANSI SGR parsing for Ratatui color spans because color/style state is a TUI rendering concern.
+- Terminal text parsing is split by responsibility. `src/terminal_text.rs` provides shared plain-text terminal rewrite handling for backend snapshots. `src/tui_terminal.rs` provides styled ANSI SGR parsing for Ratatui color spans, with parser regression tests in `src/tui_terminal_tests.rs`.
 - TUI theme selection is isolated in `src/tui_theme.rs`. The CLI accepts `dark`, `light`, or `system`, with `system` querying the terminal background before raw mode and falling back to dark for unsupported terminals or non-interactive smoke runs.
 - This separation keeps SOLID boundaries: backend runtime has one reason to change, client transport has one reason to change, TUI state/rendering has one reason to change, and the binary only wires them together.
 - New TUI features should add typed wrappers to `BackendClient` before adding UI controls. Avoid calling built-in backend internals from `tui.rs` or the binary.
