+++ src/lib.rs
@@ -1,5 +1,8 @@
 mod protocol;
 mod terminal_text;
+mod tui_input;
+mod tui_model;
+mod tui_render;
 mod tui_terminal;
 mod tui_theme;
